Refactor: add exception name to domain exceptions
