+++ src/modules/Media/Domain/UserUploadDomainException.ts
@@ -4,7 +4,7 @@ export class UserUploadDomainException extends DomainException {
   public static invalidUserUploadIdId = 'user_upload_domain_invalid_user_upload_id'
 
   private constructor(message: string, id: string) {
-    super(message, id)
+    super(message, id, UserUploadDomainException.name)
   }
 
   public static invalidUserUploadId(userUploadId: string) {
